Rpi_Python_Rover/Rpi_Python_Azure_Function_Remote_Control/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-iot-hub
azure-iothub-service-client